==> research/08_model_register.ipynb <==
import os

%pwd

os.chdir("../")


from dataclasses import dataclass
from pathlib import Path

@dataclass(frozen=True)
class ModelRegisterConfig:
  dagshu_tracking_uri : Path
  run_json_info: Path
  repo_owner: str
  repo_name: str

from pathlib import Path

CONFIG_FILE_PATH = Path("E:\projects\Delivery-time-prediction-for-food-devlivery-industry\config\config.yaml")
PARAMS_FILE_PATH = Path("E:\projects\Delivery-time-prediction-for-food-devlivery-industry\params.yaml")
SCHEMA_FILE_PATH = Path("E:\projects\Delivery-time-prediction-for-food-devlivery-industry\schema.yaml")

from Deliveryprediction.constants import *
from Deliveryprediction.utils.common import read_yaml, create_directories

class ConfigurationManager:
    def __init__(
        self,
        config_filepath = CONFIG_FILE_PATH,
        params_filepath = PARAMS_FILE_PATH,
        schema_filepath = SCHEMA_FILE_PATH):

        self.config = read_yaml(config_filepath)
        self.params = read_yaml(params_filepath)
        self.schema = read_yaml(schema_filepath)

        create_directories([self.config.artifacts_root])


    
    def get_model_registry_config(self) -> ModelRegisterConfig:
        config = self.config.model_registry    

        #create_directories([config.root_dir])

        model_registry_config = ModelRegisterConfig(
            dagshu_tracking_uri = config.dagshu_tracking_uri,
            run_json_info = config.run_json_info,
            repo_name= config.repo_name,
            repo_owner= config.repo_owner,

        )

        return model_registry_config

import mlflow
import dagshub
import json
from pathlib import Path
from mlflow import MlflowClient
from Deliveryprediction import logger

class MLFlowModelRegistry:
    def __init__(self,logger,config=ModelRegisterConfig):
        self.config = config
        self.repo_owner = config.repo_owner
        self.repo_name = config.repo_name
        self.logger = logger
        self.tracking_uri = config.dagshu_tracking_uri
        self.run_info_path = Path(config.run_json_info)
        self._initialize_dagshub()
        self._set_tracking_uri()
        self.client = MlflowClient()
        

    def _initialize_dagshub(self):
        dagshub.init(repo_owner=self.repo_owner, repo_name=self.repo_name, mlflow=True)

    def _set_tracking_uri(self):
        mlflow.set_tracking_uri(self.tracking_uri)

    def load_model_information(self):
        with open(self.run_info_path) as f:
            return json.load(f)

    def register_and_transition_model(self):
        run_info = self.load_model_information()
        run_id = run_info["run_id"]
        model_name = run_info["model_name"]
        
        model_registry_path = f"runs:/{run_id}/{model_name}"
        model_version = mlflow.register_model(model_uri=model_registry_path, name=model_name)
        
        registered_model_version = model_version.version
        registered_model_name = model_version.name
        
        self.logger.info(f"The latest model version in model registry is {registered_model_version}")
        
        self.client.transition_model_version_stage(
            name=registered_model_name,
            version=registered_model_version,
            stage="Staging"
        )
        
        self.logger.info("Model pushed to Staging stage")


    

config=ConfigurationManager()

model_registry=config.get_model_registry_config()

model_register = MLFlowModelRegistry(logger=logger,config=model_registry)

model_register.register_and_transition_model()
